Support 64 bit integers in _ut_assert_int_equal

+++ src/ut-assert.h
@@ -25,7 +25,7 @@ void _ut_assert_is_not_error(const char *file, int line, const char *name,
 #define ut_assert_int_equal(a, b)                                              \
   _ut_assert_int_equal(__FILE__, __LINE__, #a, a, #b, b)
 void _ut_assert_int_equal(const char *file, int line, const char *a_name,
-                          int a_value, const char *b_name, int b_value);
+                          int64_t a_value, const char *b_name, int64_t b_value);
 
 /// Asserts that the two floating point numbers [a] and [b] are equal.
 #define ut_assert_float_equal(a, b)                                            \
